Updated the validfilter annotation, validator and tests.
