+++ src/main/java/se/sundsvall/educationfinder/api/validation/FilterType.java
@@ -1,0 +1,6 @@
+package se.sundsvall.educationfinder.api.validation;
+
+public enum FilterType {
+	COURSE,
+	STATISTICS
+}
